somnolencia/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyserial>=3.5
requests>=2.31.0
pygame>=2.5.0